+++ packages/button/src/presenters/stylesheet.js
@@ -18,9 +18,13 @@ function getButtonDefaultButtonRules(themeData) {
     overflow: "hidden",
     padding: `${themeData["button.verticalPadding"]}
       ${themeData["button.horizontalPadding"]}`,
+    position: "relative",
     textAlign: "center",
     textDecoration: "none",
     textOverflow: "ellipsis",
+    transitionDuration: "0.3s",
+    transitionProperty: "box-shadow",
+    transitionTimingFunction: "cubic-bezier(0.4, 0, 0.2, 1)",
     userSelect: "none",
     whiteSpace: "nowrap",
     MozOsxFontSmoothing: "grayscale",
@@ -78,7 +82,8 @@ function getButtonRulesByType(type, themeData) {
 function getButtonRulesByDisabled(type, themeData) {
   return {
     opacity: themeData["component.disabled.opacity"],
-    cursor: "default"
+    cursor: "default",
+    pointerEvents: "none"
   };
 }
 
@@ -187,8 +192,9 @@ function getButtonPressedRulesByType(type, themeData) {
   }
 }
 
-export default function stylesheet(props, themeData) {
+export default function stylesheet(props, themeData, theme) {
   const { disabled, hasFocus, hasHover, isPressed, type, width } = props;
+  const iconSize = theme.densityId === "medium-density" ? "24px" : "16px";
 
   return {
     button: {
@@ -205,11 +211,14 @@ export default function stylesheet(props, themeData) {
     },
     icon: {
       display: "inline-block",
-      position: "relative",
-      height: "24px",
-      verticalAlign: "middle",
+      position: "absolute",
+      height: iconSize,
       marginRight: themeData["button.gutter"],
-      marginBottom: "1px"
+      top: "50%",
+      transform: "translateY(-50%)"
+    },
+    iconText: {
+      marginLeft: `calc(${iconSize} + ${themeData["button.gutter"]})`
     }
   };
 }