+++ packages/dropdown/src/presenters/InputPresenter.js
@@ -1,10 +1,20 @@
 import React from "react";
 import { css } from "emotion";
-import { Caret24, CaretDownSUI } from "@hig/icons";
+import { CaretDownMUI, CaretDownSUI } from "@hig/icons";
 import Input from "@hig/input";
 import ThemeContext from "@hig/theme-context";
 
 import stylesheet from "./InputPresenter.stylesheet";
+
+function customStyles(inputStylesheet) {
+  return {
+    ...inputStylesheet,
+    input: {
+      ...inputStylesheet.input,
+      cursor: "pointer"
+    }
+  };
+}
 
 export default function InputPresenter(props) {
   const { isOpen, ...otherProps } = props;
@@ -13,18 +23,14 @@ export default function InputPresenter(props) {
     <ThemeContext.Consumer>
       {({ resolvedRoles, metadata }) => {
         const CaretIcon =
-          metadata.densityId === "medium-density" ? Caret24 : CaretDownSUI;
+          metadata.densityId === "medium-density" ? CaretDownMUI : CaretDownSUI;
         const iconColor = isOpen
           ? resolvedRoles["colorScheme.accentColor"]
           : resolvedRoles["colorScheme.iconColor"];
 
         return (
           <div className={css(stylesheet(props, resolvedRoles).wrapper)}>
-            <Input
-              {...otherProps}
-              stylesheet={stylesheet(props, resolvedRoles).input}
-              readOnly
-            />
+            <Input {...otherProps} stylesheet={customStyles} readOnly />
             <CaretIcon
               style={stylesheet(props, resolvedRoles).caret}
               color={iconColor}