+++ packages/progress-ring/src/constants.js
@@ -1,0 +1,2 @@
+export const availableSizes = Object.freeze(["xs", "s", "m", "l", "xl"]);
+export const availableSurfaces = Object.freeze([100, 200, 250, 300, 350]);