+++ packages/progress-ring/src/presenters/ProgressRingPresenter.test.js
@@ -14,6 +14,28 @@ describe("progress-ring/presenters/ProgressRingPresenter", () => {
         percentComplete: 33,
         size: "s"
       }
+    },
+    {
+      description: "renders with surface prop",
+      props: {
+        percentComplete: 33,
+        surface: 200
+      }
+    },
+    {
+      description: "renders with mask prop",
+      props: {
+        percentComplete: 33,
+        mask: "red"
+      }
+    },
+    {
+      description: "renders with both surface & mask props",
+      props: {
+        percentComplete: 33,
+        surface: 200,
+        mask: "red"
+      }
     }
   ].forEach(({ description, props: { ...otherProps } }) => {
     it(description, () => {