+++ packages/side-nav/src/__stories__/ExampleSideNav.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { action } from "@storybook/addon-actions";
 import { boolean, text } from "@storybook/addon-knobs/react";
-import { Insight24, ProductsAndServices24, Collaboration24 } from "@hig/icons";
+import { Graph24, ProductsAndServices24, Collaboration24 } from "@hig/icons";
 import Typography from "@hig/typography";
 
 import SideNav from "../index";
@@ -10,13 +10,15 @@ const ExampleSideNav = () => (
   <SideNav
     headerLabel={text("Header Label", "Storybook")}
     headerLink={text("Header Link", "https://www.autodesk.com")}
+    onClickHeader={action("onClick")}
     onMinimize={action("onMinimize")}
     showMinimizeButton={boolean("Show Minimize Button", false)}
     superHeaderLabel={text("Superheader Label", "HIG")}
     superHeaderLink={text("Superheader Link", "https://www.autodesk.com")}
+    onClickSuperHeader={action("onClick")}
     groups={
       <SideNav.Group>
-        <SideNav.Module title="Module 1" icon={<Insight24 />} activeChildren>
+        <SideNav.Module title="Module 1" icon={<Graph24 />} activeChildren>
           <SideNav.Submodule title="Submodule 1" />
           <SideNav.Submodule title="Submodule 2" active />
         </SideNav.Module>