+++ packages/banner/CHANGELOG.md
@@ -1,3 +1,18 @@
+# [@hig/banner-v2.0.3](https://github.com/Autodesk/hig/compare/@hig/banner@2.0.2...@hig/banner@2.0.3) (2019-07-11)
+
+
+### Bug Fixes
+
+* update Banner to the newest Icons version ([e91c6cd](https://github.com/Autodesk/hig/commit/e91c6cd))
+
+# [@hig/banner-v2.0.2](https://github.com/Autodesk/hig/compare/@hig/banner@2.0.1...@hig/banner@2.0.2) (2019-06-25)
+
+
+### Bug Fixes
+
+* should use icon-button for dismiss instead of icon ([9c72ac5](https://github.com/Autodesk/hig/commit/9c72ac5))
+* swap icons based on density ([0c45e4c](https://github.com/Autodesk/hig/commit/0c45e4c))
+
 # [@hig/banner-v2.0.1](https://github.com/Autodesk/hig/compare/@hig/banner@2.0.0...@hig/banner@2.0.1) (2019-03-13)