+++ packages/project-account-switcher/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@hig/project-account-switcher-v2.0.4](https://github.com/Autodesk/hig/compare/@hig/project-account-switcher@2.0.3...@hig/project-account-switcher@2.0.4) (2019-07-11)
+
+
+### Bug Fixes
+
+* update ProjectAccountSwitcher to newest Icons ([ef2cf9c](https://github.com/Autodesk/hig/commit/ef2cf9c))
+
+# [@hig/project-account-switcher-v2.0.3](https://github.com/Autodesk/hig/compare/@hig/project-account-switcher@2.0.2...@hig/project-account-switcher@2.0.3) (2019-06-12)
+
+
+### Bug Fixes
+
+* dropdown list to be sized based on text width ([670cc3b](https://github.com/Autodesk/hig/commit/670cc3b))
+
+# [@hig/project-account-switcher-v2.0.2](https://github.com/Autodesk/hig/compare/@hig/project-account-switcher@2.0.1...@hig/project-account-switcher@2.0.2) (2019-06-05)
+
+
+### Bug Fixes
+
+* project switcher not show image in dropdown ([f59668b](https://github.com/Autodesk/hig/commit/f59668b))
+
 # [@hig/project-account-switcher-v2.0.1](https://github.com/Autodesk/hig/compare/@hig/project-account-switcher@2.0.0...@hig/project-account-switcher@2.0.1) (2019-04-01)